requirements.txt
# PaddleOCR and related dependencies
paddlepaddle-gpu>=3.0.0
paddleocr>=2.7.0
paddlex[ocr]>=3.0.0

# Common dependencies for OCR projects
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.3.0
matplotlib>=3.3.0

# Web framework (optional)
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Utilities
requests>=2.25.0